Correctly working abstraction of lcd7920 to Driver.
