+++ src/Display/Menu.cpp
@@ -69,7 +69,8 @@
 
 #if SUPPORT_12864_LCD
 
-#include "ST7920/lcd7920.h"
+//#include "ST7920/lcd7920.h"
+#include "Driver.h"
 #include "RepRap.h"
 #include "Platform.h"
 #include "Display/Display.h"
@@ -81,7 +82,7 @@
 const uint32_t InactivityTimeout = 20000;		// inactivity timeout
 const uint32_t ErrorTimeout = 6000;				// how long we display an error message for
 
-Menu::Menu(Lcd7920& refLcd) noexcept
+Menu::Menu(Driver& refLcd) noexcept
 	: lcd(refLcd),
 	  timeoutValue(0), lastActionTime(0),
 	  selectableItems(nullptr), unSelectableItems(nullptr), highlightedItem(nullptr), numNestedMenus(0),
