+++ src/Display/MenuItem.h
@@ -13,7 +13,8 @@
 #if SUPPORT_12864_LCD
 
 #include "General/FreelistManager.h"
-#include "ST7920/lcd7920.h"
+//#include "ST7920/lcd7920.h"
+#include "Driver.h"
 #include "Storage/MassStorage.h"
 
 // Menu item class hierarchy
@@ -28,7 +29,7 @@ public:
 	static constexpr Visibility AlwaysVisible = 0;
 
 	// Draw this element on the LCD respecting 'maxWidth' and 'highlight'
-	virtual void Draw(Lcd7920& lcd, PixelNumber maxWidth, bool highlight, PixelNumber tOffset) noexcept = 0;
+	virtual void Draw(Driver& lcd, PixelNumber maxWidth, bool highlight, PixelNumber tOffset) noexcept = 0;
 
 	// Select this element with a push of the encoder.
 	// If it returns nullptr false go into adjustment mode, if we can adjust the item.
@@ -50,7 +51,7 @@ public:
 	virtual bool Adjust(int clicks) noexcept { return true; }
 
 	// If the width was specified as zero, update it with the actual width. Also update the height.
-	virtual void UpdateWidthAndHeight(Lcd7920& lcd) noexcept = 0;
+	virtual void UpdateWidthAndHeight(Driver& lcd) noexcept = 0;
 
 	// DC: I don't know what this one is for, the person who wrote it didn't document it
 	virtual PixelNumber GetVisibilityRowOffset(PixelNumber tCurrentOffset, PixelNumber fontHeight) const noexcept { return 0; }
@@ -63,7 +64,7 @@ public:
 	bool IsVisible() const noexcept;
 
 	// Erase this item if it is drawn but should not be visible
-	void EraseIfInvisible(Lcd7920& lcd, PixelNumber tOffset) noexcept;
+	void EraseIfInvisible(Driver& lcd, PixelNumber tOffset) noexcept;
 
 	// Return the width of this item in pixels
 	PixelNumber GetWidth() const noexcept { return width; }
@@ -76,10 +77,10 @@ protected:
 
 	// Print the item starting at the current cursor position, which may be off screen. Used to find the width and also to really print the item.
 	// Overridden for items that support variable alignment
-	virtual void CorePrint(Lcd7920& lcd) noexcept { }
+	virtual void CorePrint(Driver& lcd) noexcept { }
 
 	// Print the item at the correct place with the correct alignment
-	void PrintAligned(Lcd7920& lcd, PixelNumber tOffset, PixelNumber rightMargin) noexcept;
+	void PrintAligned(Driver& lcd, PixelNumber tOffset, PixelNumber rightMargin) noexcept;
 
 	const PixelNumber row, column;
 	PixelNumber width, height;
@@ -102,11 +103,11 @@ public:
 	void operator delete(void* p) noexcept { FreelistManager::Release<TextMenuItem>(p); }
 
 	TextMenuItem(PixelNumber r, PixelNumber c, PixelNumber w, Alignment a, FontNumber fn, Visibility vis, const char *t) noexcept;
-	void Draw(Lcd7920& lcd, PixelNumber maxWidth, bool highlight, PixelNumber tOffset) noexcept override;
-	void UpdateWidthAndHeight(Lcd7920& lcd) noexcept override;
+	void Draw(Driver& lcd, PixelNumber maxWidth, bool highlight, PixelNumber tOffset) noexcept override;
+	void UpdateWidthAndHeight(Driver& lcd) noexcept override;
 
 protected:
-	void CorePrint(Lcd7920& lcd) noexcept override;
+	void CorePrint(Driver& lcd) noexcept override;
 
 private:
 	const char *text;
@@ -119,14 +120,14 @@ public:
 	void operator delete(void* p) noexcept { FreelistManager::Release<ButtonMenuItem>(p); }
 
 	ButtonMenuItem(PixelNumber r, PixelNumber c, PixelNumber w, FontNumber fn, Visibility vis, const char *t, const char *cmd, const char *acFile) noexcept;
-	void Draw(Lcd7920& lcd, PixelNumber maxWidth, bool highlight, PixelNumber tOffset) noexcept override;
-	void UpdateWidthAndHeight(Lcd7920& lcd) noexcept override;
+	void Draw(Driver& lcd, PixelNumber maxWidth, bool highlight, PixelNumber tOffset) noexcept override;
+	void UpdateWidthAndHeight(Driver& lcd) noexcept override;
 	bool Select(const StringRef& cmd) noexcept override;
 
 	PixelNumber GetVisibilityRowOffset(PixelNumber tCurrentOffset, PixelNumber fontHeight) const noexcept override;
 
 protected:
-	void CorePrint(Lcd7920& lcd) noexcept override;
+	void CorePrint(Driver& lcd) noexcept override;
 
 private:
 	const char *text;
@@ -141,18 +142,18 @@ public:
 	void operator delete(void* p) noexcept { FreelistManager::Release<ValueMenuItem>(p); }
 
 	ValueMenuItem(PixelNumber r, PixelNumber c, PixelNumber w, Alignment a, FontNumber fn, Visibility vis, bool adj, unsigned int v, unsigned int d) noexcept;
-	void Draw(Lcd7920& lcd, PixelNumber maxWidth, bool highlight, PixelNumber tOffset) noexcept override;
+	void Draw(Driver& lcd, PixelNumber maxWidth, bool highlight, PixelNumber tOffset) noexcept override;
 	bool Select(const StringRef& cmd) noexcept override;
 	bool CanAdjust() const noexcept override { return true; }
 	bool Adjust(int clicks) noexcept override;
-	void UpdateWidthAndHeight(Lcd7920& lcd) noexcept override;
+	void UpdateWidthAndHeight(Driver& lcd) noexcept override;
 
 	PixelNumber GetVisibilityRowOffset(PixelNumber tCurrentOffset, PixelNumber fontHeight) const noexcept override;
 
 	unsigned int GetReferencedToolNumber() const noexcept;
 
 protected:
-	void CorePrint(Lcd7920& lcd) noexcept override;
+	void CorePrint(Driver& lcd) noexcept override;
 
 private:
 	enum class AdjustMode : uint8_t { displaying, adjusting, liveAdjusting };
@@ -188,11 +189,11 @@ public:
 	void operator delete(void* p) noexcept { FreelistManager::Release<FilesMenuItem>(p); }
 
 	FilesMenuItem(PixelNumber r, PixelNumber c, PixelNumber w, FontNumber fn, Visibility vis, const char *cmd, const char *dir, const char *acFile, unsigned int nf) noexcept;
-	void Draw(Lcd7920& lcd, PixelNumber rightMargin, bool highlight, PixelNumber tOffset) noexcept override;
+	void Draw(Driver& lcd, PixelNumber rightMargin, bool highlight, PixelNumber tOffset) noexcept override;
 	void Enter(bool bForwardDirection) noexcept override;
 	int Advance(int nCounts) noexcept override;
 	bool Select(const StringRef& cmd) noexcept override;
-	void UpdateWidthAndHeight(Lcd7920& lcd) noexcept override;
+	void UpdateWidthAndHeight(Driver& lcd) noexcept override;
 
 	PixelNumber GetVisibilityRowOffset(PixelNumber tCurrentOffset, PixelNumber fontHeight) const noexcept override;
 
@@ -202,7 +203,7 @@ protected:
 	void vResetViewState() noexcept;
 
 private:
-	void ListFiles(Lcd7920& lcd, PixelNumber rightMargin, bool highlight, PixelNumber tOffset) noexcept;
+	void ListFiles(Driver& lcd, PixelNumber rightMargin, bool highlight, PixelNumber tOffset) noexcept;
 	uint8_t GetDirectoryNesting() const noexcept;
 
 	const unsigned int numDisplayLines;
@@ -236,8 +237,8 @@ public:
 
 	ImageMenuItem(PixelNumber r, PixelNumber c, Visibility vis, const char *pFileName) noexcept;
 
-	void Draw(Lcd7920& lcd, PixelNumber rightMargin, bool highlight, PixelNumber tOffset) noexcept override;
-	void UpdateWidthAndHeight(Lcd7920& lcd) noexcept override;
+	void Draw(Driver& lcd, PixelNumber rightMargin, bool highlight, PixelNumber tOffset) noexcept override;
+	void UpdateWidthAndHeight(Driver& lcd) noexcept override;
 
 private:
 	String<MaxFilenameLength> fileName;
